requirements.txt
freetype-py
imageio
networkx
numpy
Pillow
pyglet==1.4.0a1
PyOpenGL
PyOpenGL_accelerate
six
trimesh
sphinx
sphinx_rtd_theme
sphinx-automodapi

